requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
chardet==2.2.1
click==6.7
colorama==0.2.5
cryptography==2.3
dominate==2.3.1
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
html5lib==0.999
idna==2.7
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.0
mimerender==0.6.0
Pillow==5.2.0
pyasn1==0.4.4
pycparser==2.18
python-dateutil==2.7.3
python-mimeparse==1.6.0
requests==2.2.1
six==1.5.2
SQLAlchemy==1.2.10
urllib3==1.7.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
